Time: 30 ms (40.13%), Space: 18 MB (20.98%) - LeetHub

+++ 268-missing-number/268-missing-number.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        long long sum=0;
-        long long size=nums.size();
-        long long req=size*(size+1)/2;
+        int sum=0;
+        int size=nums.size();
+        int req=size*(size+1)/2;
         for(int i=0;i<size;i++)
             sum+=nums[i];
         return req-sum;
